requirements.txt
SQLAlchemy~=2.0.41
aiogram~=3.3.0
APScheduler~=3.11.0
alembic~=1.16.2
python-dotenv~=0.9.1
aiohttp~=3.9.5
gspread~=6.2.1
oauth2client~=4.1.3
google-api-python-client~=2.175.0
aiosqlite==0.21.0